Split the flake into overlays and modules.
Having a flake which serves as an index is nicer for people who are
trying to understand what are the entry points.

Also, splitting the flake into an overlay and a module is better as this
removes the potential for having multiple Nixpkgs evaluations.
Previously we would have a Nixpkgs evaluation for the inventree
commands, and at least another one for NixOS evaluation.