Prefer python packages provided by Nixpkgs.
pip2nix wants to manage the world, however, as Nixpkgs version increase, the
number of packages which would be handled by Nixpkgs would increase as well.

This patch will select the packages which have the newest version, from either
the Nixpkgs version or from the pip2nix collection. This should reduce the
amount of recompilation of python packages which has to be done before
installing inventree, and maximize the reusability of existing packages over
time.